+++ TreeTycoon/Engine/Inc/Entity.hpp
@@ -38,7 +38,7 @@ public:
 
 	int8_t getUpdatePriority() const final
 	{
-		return ConvertTo<int16_t>( UpdatePriority::Entity );
+		return ConvertTo<int8_t>( UpdatePriority::Entity );
 	}
 
 	Status getStatus() const;
