+++ TreeTycoon/Engine/Inc/Renderer.hpp
@@ -16,7 +16,7 @@ class RendererClass final :
 public:
 	int8_t getUpdatePriority() const override
 	{
-		return ConvertTo<int16_t>( UpdatePriority::Renderer );
+		return ConvertTo<int8_t>( UpdatePriority::Renderer );
 	}
 
 	void add( IDrawable* drawable );
